requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asyncio==3.4.3
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
instructor==1.4.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.6
openai==1.45.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
solidity-parser==0.1.1
tenacity==8.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
cohere==4.1.0
langchain_experimental
bs4
markdown2
tiktoken
pinecone
